+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Go and NodeJS output formats
 - Exclude flag to ignore specific directories
 
 ### Changed